feat: implement database reset functionality and update scripts
