+++ package.json
@@ -10,10 +10,10 @@
     "db:latest": "kysely migrate:latest",
     "db:up": "kysely migrate:up",
     "db:down": "kysely migrate:down",
-    "db:drop": "tsx ./src/scripts/drop_database.ts",
-    "db:create": "tsx ./src/scripts/create_database.ts",
+    "db:reset": "tsx ./scripts/reset_database.ts",
+    "db:create": "tsx ./scripts/create_database.ts",
     "format": "biome format --write ./src",
-    "migrate": "tsx ./src/scripts/migrate.ts"
+    "migrate": "tsx ./scripts/migrate.ts"
   },
   "dependencies": {
     "@bugsnag/cuid": "^3.1.1",
