+++ scripts/data/resources.json
